# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BIMLog API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{projectId}/rfis:
    get:
      operationId: listRfis
      tags: [rfis]
      summary: List RFIs for a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RFI list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rfi"
    post:
      operationId: createRfi
      tags: [rfis]
      summary: Create an RFI
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRfiRequest"
      responses:
        "201":
          description: RFI created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfi"
  /projects/{projectId}/rfis/{rfiId}:
    patch:
      operationId: updateRfi
      tags: [rfis]
      summary: Update an RFI
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: rfiId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRfiRequest"
      responses:
        "200":
          description: RFI updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfi"
...
  /projects/{projectId}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List team members for a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
    post:
      operationId: addMember
      tags: [members]
      summary: Add a team member to a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
  /projects/{projectId}/members/{memberId}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a team member's role
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
    delete:
      operationId: removeMember
      tags: [members]
      summary: Remove a team member from a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        code:
          type: string
        status:
          type: string
        createdById:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        fileCount:
          type: integer
      required:
        - id
        - name
        - code
        - status
        - createdById
        - createdAt
        - updatedAt
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        code:
          type: string
      required:
        - name
        - code
...
    Rfi:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        number:
          type: string
        subject:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        createdById:
          type: integer
        createdByName:
          type: string
        dueDate:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time
        response:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - number
        - subject
        - status
        - priority
        - createdById
        - createdByName
        - createdAt
        - updatedAt
    CreateRfiRequest:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedToId:
          type: integer
        dueDate:
          type: string
          format: date-time
      required:
        - subject
        - priority
    UpdateRfiRequest:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_review, responded, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedToId:
          type: integer
        response:
          type: string
        dueDate:
          type: string
          format: date-time
...
    ProjectMember:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: integer
        userFullName:
          type: string
        userEmail:
          type: string
        userCompanyName:
          type: string
        role:
          type: string
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - userId
        - userFullName
        - userEmail
        - role
        - joinedAt
    AddMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [project_admin, company_lead, drafter, project_manager, read_only]
      required:
        - email
        - role
    UpdateMemberRequest:
      type: object
      properties:
        role:
          type: string
          enum: [project_admin, company_lead, drafter, project_manager, read_only]
      required:
        - role