# lib/api-zod/src/generated/types/rfi.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIMLog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Rfi {
  id: number;
  projectId: number;
  number: string;
  subject: string;
  description?: string;
  status: string;
  priority: string;
  assignedToId?: number;
  assignedToName?: string;
  createdById: number;
  createdByName: string;
  dueDate?: Date;
  respondedAt?: Date;
  response?: string;
  createdAt: Date;
  updatedAt: Date;
}